+++ database/migrations/tabantaj2/2021_07_06_222220_create_documentos_table.php
@@ -25,10 +25,10 @@ class CreateDocumentosTable extends Migration
             $table->string('version');
             $table->dateTime('fecha');
             $table->string('archivo');
-            $table->unsignedBigInteger('elaboro_id')->nullable();
-            $table->unsignedBigInteger('reviso_id')->nullable();
-            $table->unsignedBigInteger('aprobo_id')->nullable();
-            $table->unsignedBigInteger('responsable_id')->nullable();
+            $table->integer('elaboro_id')->nullable();
+            $table->integer('reviso_id')->nullable();
+            $table->integer('aprobo_id')->nullable();
+            $table->integer('responsable_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
