+++ database/migrations/tabantaj2/2021_07_12_200050_add_foreign_key_to_indicadores_sgsis_table.php
@@ -16,7 +16,7 @@ class AddForeignKeyToIndicadoresSgsisTable extends Migration
         Schema::table('indicadores_sgsis', function (Blueprint $table) {
             $table->unsignedInteger('id_proceso')->nullable();
             $table->foreign('id_proceso')->references('id')->on('procesos');
-            $table->unsignedBigInteger('id_empleado')->nullable();
+            $table->integer('id_empleado')->nullable();
         });
 
         Schema::table('indicadores_sgsis', function ($table) {
