+++ database/migrations/tabantaj2/2021_07_20_224224_create_analisis_de_riesgo_table.php
@@ -19,7 +19,7 @@ class CreateAnalisisDeRiesgoTable extends Migration
             $table->string('tipo');
             $table->date('fecha');
             $table->string('porcentaje_implementacion');
-            $table->unsignedBigInteger('id_empleado')->nullable();
+            $table->integer('id_empleado')->nullable();
             $table->foreign('id_empleado')->references('id')->on('empleados');
             $table->integer('estatus');
             $table->timestamps();
