+++ database/migrations/tabantaj2/2021_08_05_181401_create_media_table.php
@@ -18,7 +18,7 @@ class CreateMediaTable extends Migration
             $table->string('mime_type')->nullable();
             $table->string('disk');
             $table->string('conversions_disk')->nullable();
-            $table->unsignedBigInteger('size');
+            $table->integer('size');
             $table->json('manipulations');
             $table->json('custom_properties');
             $table->json('responsive_images');
