+++ database/migrations/tabantaj2/2021_08_10_223433_create_evidencias_quejas_table.php
@@ -15,7 +15,7 @@ class CreateEvidenciasQuejasTable extends Migration
     {
         Schema::create('evidencias_quejas', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('id_quejas')->nullable();
+            $table->integer('id_quejas')->nullable();
             $table->string('evidencia');
             $table->timestamps();
             $table->softDeletes();
