+++ database/migrations/tabantaj2/2021_08_11_204548_create_plan_implementacionables_table.php
@@ -15,8 +15,8 @@ class CreatePlanImplementacionablesTable extends Migration
     {
         Schema::create('plan_implementacionables', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('plan_implementacion_id')->nullable();
-            $table->unsignedBigInteger('plan_implementacionable_id')->nullable();
+            $table->integer('plan_implementacion_id')->nullable();
+            $table->integer('plan_implementacionable_id')->nullable();
             $table->string('plan_implementacionable_type')->nullable();
             $table->foreign('plan_implementacion_id')->references('id')->on('plan_implementacions')->onDelete('cascade');
             $table->timestamps();
