+++ database/migrations/tabantaj2/2021_08_27_184319_add_responsable_id_to_minutasaltadireccions_table.php
@@ -14,7 +14,7 @@ class AddResponsableIdToMinutasaltadireccionsTable extends Migration
     public function up()
     {
         Schema::table('minutasaltadireccions', function (Blueprint $table) {
-            $table->unsignedBigInteger('responsable_id');
+            $table->integer('responsable_id');
             $table->foreign('responsable_id')->references('id')->on('empleados');
         });
     }
