+++ database/migrations/tabantaj2/2021_08_27_190805_create_actividades_riesgos_table.php
@@ -15,7 +15,7 @@ class CreateActividadesRiesgosTable extends Migration
     {
         Schema::create('actividades_riesgos', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('riesgo_id');
+            $table->integer('riesgo_id');
             $table->string('actividad');
             $table->dateTime('fecha_inicio');
             $table->dateTime('fecha_fin');
