+++ database/migrations/tabantaj2/2021_09_10_184354_add_fk_to_auditoria_internas_table.php
@@ -14,7 +14,7 @@ class AddFkToAuditoriaInternasTable extends Migration
     public function up()
     {
         Schema::table('auditoria_internas', function (Blueprint $table) {
-            $table->unsignedBigInteger('lider_id');
+            $table->integer('lider_id');
             $table->foreign('lider_id')->references('id')->on('empleados')->onUpdate('cascade')->onDelete('cascade');
         });
     }
