+++ database/migrations/tabantaj2/2021_09_10_184758_create_auditoria_interno_empleado_table.php
@@ -15,7 +15,7 @@ class CreateAuditoriaInternoEmpleadoTable extends Migration
     {
         Schema::create('auditoria_interno_empleado', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('empleado_id');
+            $table->integer('empleado_id');
             $table->foreign('empleado_id')->references('id')->on('empleados')->onUpdate('cascade')->onDelete('cascade');
             $table->unsignedInteger('auditoria_id');
             $table->foreign('auditoria_id')->references('id')->on('auditoria_internas')->onUpdate('cascade')->onDelete('cascade');
