+++ database/migrations/tabantaj2/2021_09_22_055723_add_columns_id_reviso_to_planificacion_controls_table.php
@@ -14,7 +14,7 @@ class AddColumnsIdRevisoToPlanificacionControlsTable extends Migration
     public function up()
     {
         Schema::table('planificacion_controls', function (Blueprint $table) {
-            $table->unsignedBigInteger('id_reviso')->nullable();
+            $table->integer('id_reviso')->nullable();
             $table->foreign('id_reviso')->references('id')->on('empleados');
         });
     }
