+++ database/migrations/tabantaj2/2021_10_14_194701_create_analisis_brechas_table.php
@@ -18,7 +18,7 @@ class CreateAnalisisBrechasTable extends Migration
             $table->string('nombre');
             $table->date('fecha');
             $table->string('porcentaje_implementacion');
-            $table->unsignedBigInteger('id_elaboro')->nullable();
+            $table->integer('id_elaboro')->nullable();
             $table->foreign('id_elaboro')->references('id')->on('empleados');
             $table->integer('estatus');
             $table->timestamps();
